% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{example.hg19.ref.cts.se}
\alias{example.hg19.ref.cts.se}
\title{Subset of hg19.ref.cts.se as a lightweighted example}
\format{
An object of class \code{SummarizedExperiment}.
}
\usage{
data(example.hg19.ref.cts.se)
}
\description{
Subset of hg19.ref.cts.se as a lightweighted example, only
contains the blood cell types and prostate.
}
\details{
All coordinates are in hg19.
}
\keyword{datasets}
