% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg19.cpg.coords}
\alias{hg19.cpg.coords}
\title{Genomic information of CpG positions in hg19}
\format{
An object of class \code{GRanges}.
}
\usage{
data(hg19.cpg.coords)
}
\description{
This dataset represents a GRanges object that contains the collection
of CpGs in chr1-chr22, chrX and chrY. Each row is a CpG. Information was extracted
from \code{BSgenome.Hsapiens.UCSC.hg19} of package \code{BSgenome}.
}
\details{
All coordinates are in hg19.
}
\examples{
  data(hg19.cpg.coords)
  hg19.cpg.coords
}
\keyword{datasets}
