% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg38.ref.anc.se}
\alias{hg38.ref.anc.se}
\title{Hg38 genomic information of cell-type-specific marker CpGs inferred from DNA
methylation atlas published by Moss 2018 Nat. Commun.
(https://www.nature.com/articles/s41467-018-07466-6).}
\format{
An object of class \code{SummarizedExperiment}.

An object of class \code{SummarizedExperiment}.
}
\usage{
data(hg38.ref.anc.se)
}
\description{
Same as \code{data(hg19.ref.anc.se)} but lifted over to hg38.
}
\details{
All coordinates are in hg38.
}
\examples{
data(hg38.ref.anc.se)
hg38.ref.anc.se
}
\keyword{datasets}
