% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pdx.counts.cts.se}
\alias{pdx.counts.cts.se}
\title{MeDIP-seq read counts on reference cell type-specific CpGs of 3 PDX samples from Berchuck et al. 2022}
\format{
An object of class \code{SummarizedExperiment}.
}
\usage{
data(pdx.counts.cts.se)
}
\description{
This dataset represents a \code{SummarizedExperiment} object that contains
MeDIP-seq read counts on reference cell type-specific CpGs of 3 PDX samples from Berchuck
et al. 2022. Each row is a CpG.
}
\details{
All coordinates are in hg19.
}
\examples{
data(pdx.counts.cts.se)
pdx.counts.cts.se
}
\keyword{datasets}
