% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decemedip-package.R
\docType{package}
\name{decemedip-package}
\alias{decemedip-package}
\title{The 'decemedip' package.}
\description{
The R package decemedip is a novel computational paradigm developed
for inferring the relative abundances of cell types and tissues from tissue bulk
or circulating cell-free DNA (cfDNA) measure by methylated DNA immunoprecipitation
sequencing (MeDIP-Seq). This paradigm allows using reference data from other
technologies such as microarray or WGBS.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Ning Shen \email{ning.shen.wk@gmail.com} (\href{https://orcid.org/0000-0002-2974-1086}{ORCID})

}
