% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decemedip.R
\name{decemedip}
\alias{decemedip}
\title{Perform cell type deconvolution with MeDIP-seq data}
\usage{
decemedip(
  sample_bam_file = NULL,
  paired_end = NULL,
  counts_cts = c(),
  counts_anc = c(),
  ref_cts = hg19.ref.cts.se,
  ref_anc = hg19.ref.anc.se,
  diagnostics = TRUE,
  seed = 2024,
  cores = 4,
  chains = 4,
  iter = 1000,
  stan_input_params = list(alpha = rep(1, SummarizedExperiment::ncol(ref_cts)), s_mu = 1,
    s_sigma = 1, n_knot_z = 0, degree_z = 3, Xi =
    cor(as.matrix(SummarizedExperiment::assays(ref_cts)[[1]])), s_theta = 1, s_tau = 1),
  stan_control = NULL,
  ...
)
}
\arguments{
\item{sample_bam_file}{A string value that indicates the file path to the bam file of a
MeDIP-seq sample of interest. If \code{sample_bam_file} is specified, please do not
specify \code{counts_cts} and \code{counts_anc} to avoid conflict.}

\item{paired_end}{A logic value that indicates whether the bam file is from paired-end
reads or single-end. Specify \code{TRUE} for paired-end and \code{FALSE} for single-end.}

\item{counts_cts}{An atomic vector of integer values that indicates the read counts of
a MeDIP-seq sample on reference sites/regions. If \code{counts_cts} and \code{counts_anc}
is specified, please do not specify \code{sample_bam_file} and \code{paired_end} to avoid conflict.}

\item{counts_anc}{An atomic vector of integer values that indicates the read counts of
a MeDIP-seq sample on reference sites/regions. If  \code{counts_cts} and \code{counts_anc}
is specified, please do not specify \code{sample_bam_file} and \code{paired_end} to avoid conflict.}

\item{ref_cts}{A \code{SummarizedExperiment} object that contains the genomic coordinates
and beta values of the cell type-specific sites/regions from reference cell types. The default
reference is explained in the manuscript of \pkg{decemedip}.}

\item{ref_anc}{A \code{SummarizedExperiment} object that contains the genomic coordinates
and beta values of the anchor sites/regions from reference cell types.  The default
reference is explained in the manuscript of \pkg{decemedip}.}

\item{diagnostics}{A logic value that indicates whether to include components of the stan model
in the output that are necessary for future diagnostics of the model, such as posterior
predictive checks. For details, please refer to the function \code{\link{plotDiagnostics}}.}

\item{seed}{The seed for random number generation in MCMC sampling.}

\item{cores}{A positive integer specifying the number of cores that can be used for
MCMC sampling. The default is 1000.}

\item{chains}{A positive integer specifying the number of Markov chains. The default is 4.}

\item{iter}{A positive integer specifying the number of iterations for each chain (including
warmup). The default is 1000.}

\item{stan_input_params}{A named list of parameters that specifies the prior of the
decemedip model.}

\item{stan_control}{A named list of parameters to control the sampler's behavior in
Stan. See the details in the documentation for the control argument in \code{\link[rstan]{stan}}.}

\item{...}{Other parameters that can be passed to the \code{\link[rstan]{sampling}} function.}
}
\description{
Perform cell type deconvolution with MeDIP-seq data
}
