% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryOnPi.R
\name{getSummaryOnPi}
\alias{getSummaryOnPi}
\title{Title}
\usage{
getSummaryOnPi(
  fit,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  digits_summary = 5
)
}
\arguments{
\item{fit}{TODO}

\item{probs}{TODO}

\item{digits_summary}{TODO}
}
\description{
Title
}
